pyproject.toml
[project]
name = "agent-task-framework"
version = "0.1.0"
description = "A sophisticated tool for transforming ambiguous user requests into actionable instructions for AI agents"
authors = [
    {name = "Agent Task Framework Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dspy-ai==2.4.4",
    "ollama==0.2.1",
    "pytest",
    "pytest-cov",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["atf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=atf --cov-report=term-missing"

[tool.coverage.run]
source = ["atf"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 